+++ eo-runtime/src/main/eo/org/eolang/regex.eo
@@ -1,4 +1,4 @@
-+package org.eolang.txt
++package org.eolang
 +rt jvm org.eolang:eo-runtime:0.1.18
 
 # Regular expression in Perl format, for example: /[a-z]+/i
