+++ eo-runtime/src/test/eo/org/eolang/regex-tests.eo
@@ -1,5 +1,4 @@
-+package org.eolang.txt
-+alias regex org.eolang.txt.regex
++package org.eolang
 +junit
 
 [] > matches-string-against-pattern
